refactor: move event handling to a separate file
This is a first step in cleaning up the supervisor code.
